Personal Advisory
Truth is.. If I WANT to be a victim, I will be. And unfortunately, you must accept being the cause if I say it is so. And it simply can't be disputed, unless of course you want victimize AND victim shame.... That'll end careers, best you just bend the knee to my desire to be offended.

-I'm joking. However, reality is not joking, and that's the form it has taken.  Be careful out their friends.